DA381A-Project submission
This commit features the accumulative works for Ramin, Oscar and Sam as per the requirements outlined in the project description. Work was completed on Sam's local system through intensive pair programming and joint collaborative efforts. Therefore every member contributed towards the success of the project, especially where refactoring or troubleshooting was concerned. It is for that reason also impossible to give a detailed list of features-per-person as all work was done in team sessions.

Co-Authored-By: Ramin <83227587+RammeDon@users.noreply.github.com>
Co-Authored-By: Oscar Karlsson <48997798+OsvarK@users.noreply.github.com>